Updates config, adds kibana (old version), tests
Moved all tests to tests_old and created file structure for new tests in tests
folder. The old tests can be moved, if they work and should be restructured so that
tests are organized in modules and classes.

Changes the config. Now the module singleton is used.

Adds the kibana dashboard functions. This are the old functions and have to be reviewed
and probably changed