Functions from statistics, import metric
Uses the functions from the statistics module whenever possible (functions in python 3.6).
The metric for spike_ref_metric or spike_height_metric can now be an importable function,
so if you enter "statistics.harmonic_mean" the harmonic_mean function from the statistics
module is used even if there is no explicit mapping for this function.