Fixes rules_loader
The RulesLoader and the FileRules loader are fixed. They now take the --rule command line argument
into account when loading rules and getting rule hashes.

The callstack of the methods to load rules is now a bit easier.

Rule hashes are now integers to prevent messing around with bytes and strings in python3. int
is choosen because the python object hash() method also uses ints.